fix(port): convergence round 1 - client/recipe/fluid/registration
Fix compile errors after the full port: client model baking (IUnbakedGeometry.bake
5-arg, BlockElementFace record, FaceBakery), screens (renderBackground/mouseScrolled
signatures), book rendering, TemplateLevel new abstracts, NBT->CUSTOM_DATA; recipe
RecipeInput/RecipeHolder/codec consistency; fluid component edges; registration
DeferredHolder generics + IdAwareObject.getId() contract.

+++ src/main/java/slimeknights/mantle/client/render/FluidRenderer.java
@@ -137,7 +137,7 @@ public class FluidRenderer {
 
     // if rotating by 90 or 270, swap U and V
     float minU, maxU, minV, maxV;
-    double size = flowing ? 8 : 16;
+    float size = flowing ? 8 : 16;
     if ((rotation % 180) == 90) {
       minU = sprite.getU(v1 * size);
       maxU = sprite.getU(v2 * size);
